Add files via upload
The "RedBus Data Scraping and Filtering with Streamlit Application" is designed to collect, analyze, and visualize bus travel data from RedBus. The project utilizes Selenium for web scraping, MySQL for storing data and Streamlit for building a user-friendly interface that enables users to filter, explore, and visualize bus travel information. This application automates the extraction of crucial data, such as bus routes, schedules, prices, and seat availability, and provides an intuitive dashboard for data analysis and decision-making.